updated delete route for logged user

+++ controllers/blogs.js
@@ -1,6 +1,15 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+// const getTokenFrom = request => {
+// 	const authorization = request.get('authorization')
+// 	if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+// 		return authorization.substring(7)
+// 	}
+// 	return null
+// }
 
 blogsRouter.get('/', async (request, response) => {
 	const blogs = await Blog
@@ -17,21 +26,18 @@ blogsRouter.get('/:id', async (request, response) => {
 	} else {
 		response.status(404).end()
 	}
-	// Blog.findById(request.params.id)
-	// 	.then(blog => {
-	// 		if(blog) {
-	// 			response.json(blog)
-	// 		} else {
-	// 			response.status(404).end()
-	// 		}
-	// 	})
 })
 
 blogsRouter.post('/', async (request, response) => {
 	const body = request.body
+	console.log('request', request.token)
+	// const token = getTokenFrom(request)
+	// const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	// if(!decodedToken.id) {
+	// 	return response.status(401).json({ error: 'token missing or invalid' })
+	// }
 
-	const user = await User.findById(body.userId)
-
+	const user = await User.findById(request.user.id)
 	const blog = new Blog({
 		title: body.title,
 		author: body.author,
@@ -53,14 +59,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-	const blogToDelete = await Blog.findByIdAndRemove(request.params.id)
-	console.log('delete', blogToDelete)
-	response.status(204).end()
-	// Blog.findByIdAndRemove(request.params.id)
-	// 	.then(() => {
-	// 		response.status(204).end()
-	// 	})
-	// 	.catch(error => next(error))
+	const blog = await Blog.findById(request.params.id)
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+	console.log('decodedtoken', decodedToken)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+
+	const user = await User.findById(decodedToken.id)
+
+	console.log('user', user.id.toString())
+	console.log('blog', blog.user.toString())
+	if (blog.user.toString() === user.id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id)
+		response.status(204).end()
+	} else {
+		response.status(401).json({ error: 'User unauthorized' })
+	}
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
